Consolidated game editor tool - Make the package structure more organized
