+++ src/main/kotlin/com/vandenbreemen/videogameusr/tools/interactor/SpriteCodeGenerationInteractor.kt
@@ -1,4 +1,4 @@
-package com.vandenbreemen.com.vandenbreemen.videogameusr.tools
+package com.vandenbreemen.videogameusr.tools.interactor
 
 import com.vandenbreemen.com.vandenbreemen.videogameusr.log.klog
 import com.vandenbreemen.viddisplayrast.data.GameDataRequirements
