+++ src/main/kotlin/com/vandenbreemen/videogameusr/tools/model/GameDataEditorModel.kt
@@ -3,11 +3,11 @@ package com.vandenbreemen.videogameusr.tools.model
 import androidx.compose.ui.graphics.Color
 import com.vandenbreemen.com.vandenbreemen.videogameusr.log.klog
 import com.vandenbreemen.com.vandenbreemen.videogameusr.model.ColorInteractor
-import com.vandenbreemen.com.vandenbreemen.videogameusr.tools.SpriteCodeGenerationInteractor
 import com.vandenbreemen.viddisplayrast.data.ByteColorDataInteractor
 import com.vandenbreemen.viddisplayrast.data.GameDataRequirements
 import com.vandenbreemen.viddisplayrast.game.Runner
 import com.vandenbreemen.videogameusr.model.game.TileBasedGameWorld
+import com.vandenbreemen.videogameusr.tools.interactor.SpriteCodeGenerationInteractor
 
 class GameDataEditorModel(private val requirements: GameDataRequirements,
                           private val tileBasedGameWorld: TileBasedGameWorld,
